// src/common/enum.ts
export enum LANG {
  ZH = 'zh',
  EN = 'en',
}

export enum TYPE {
  PureObject = '[object Object]',
  Boolean = '[object Boolean]',
  Function = '[object Function]',
  Array = '[object Array]',
  Number = '[object Number]',
}

export enum SELECT_TYPE {
  SINGLE = 'single',
  MULTIPLE = 'multiple',
  RANGE = 'range',
}

export enum PANEL_TYPE {
  WEEK = 'week',
  MONTH = 'month',
}

export enum DATE_UNIT {
  DAY = 'day',
  WEEK = 'week',
  MONTH = 'month',
  YEAR = 'year',
}